Consolidate patch(es) into existing files
